+++ src/main/kotlin/com/github/intellij/plugins/mt4ij/listeners/VFSListener.kt
@@ -1,15 +1,12 @@
 package com.github.intellij.plugins.mt4ij.listeners;
 
+import com.github.intellij.plugins.mt4ij.ApiHelpers.Companion.getActiveProject
+import com.github.intellij.plugins.mt4ij.ApiHelpers.Companion.getContentEntry
+import com.github.intellij.plugins.mt4ij.ApiHelpers.Companion.getModelForFile
+import com.github.intellij.plugins.mt4ij.ApiHelpers.Companion.invokeCommit
 import com.github.intellij.plugins.mt4ij.config.SettingsStorage
-import com.intellij.ide.DataManager
-import com.intellij.openapi.actionSystem.CommonDataKeys
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.roots.ContentEntry
-import com.intellij.openapi.roots.ModifiableRootModel
-import com.intellij.openapi.roots.ModuleRootManager
-import com.intellij.openapi.roots.ProjectFileIndex
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.newvfs.BulkFileListener
@@ -18,7 +15,6 @@ import org.jetbrains.jps.model.java.JavaSourceRootType.SOURCE
 import org.jetbrains.jps.model.java.JavaSourceRootType.TEST_SOURCE
 import java.io.File
 import java.net.URL
-import java.util.concurrent.TimeUnit
 
 /*
     // File system events listening
@@ -51,69 +47,17 @@ import java.util.concurrent.TimeUnit
  */
 
 internal class VFSListener : BulkFileListener {
-    private val GET_ACTIVE_PROJECT_TIMEOUT = 60
-    private val log : Logger               = Logger.getInstance(VFSListener::class.java)
-
-    private fun getContentEntry(model: ModifiableRootModel): ContentEntry? {
-        val contentEntries = model.contentEntries
-        return (if (contentEntries.isNotEmpty()) contentEntries[0] else null)
-    }
-
-    private fun getContentEntry(model: ModifiableRootModel, virtualFile: VirtualFile): ContentEntry? {
-        val contentEntries = model.contentEntries
-        return (if (contentEntries.isEmpty()) model.addContentEntry(virtualFile) else contentEntries[0])
-    }
-
-    private fun getModelForFile(project: Project, virtualFile: VirtualFile): ModifiableRootModel? {
-        val fileIndex = ProjectFileIndex.SERVICE.getInstance(project)
-        val module    = fileIndex.getModuleForFile(virtualFile)
-
-        if (null != module) {
-            if (module.moduleTypeName.equals("JAVA_MODULE")) {
-                return ModuleRootManager
-                    .getInstance(module)
-                    .modifiableModel
-            }
-        }
-
-        return null
-    }
-
-    private fun getActiveProject() : Project {
-        val dataContext = DataManager.getInstance().dataContextFromFocusAsync.blockingGet(GET_ACTIVE_PROJECT_TIMEOUT, TimeUnit.SECONDS)
-        val project     = dataContext?.getData(CommonDataKeys.PROJECT.name) as Project
-
-        if (null == project) {
-            log.warn("Active project not found.")
-        }
-
-        return project
-    }
-
-    private fun invokeLater(runnable: Runnable, project: Project) {
-        val application = ApplicationManager.getApplication()
-
-        application.invokeLater(
-            { application.runWriteAction(runnable) },
-            project.disposed
-        )
-    }
-
-    private fun invokeCommit(model: ModifiableRootModel, project: Project) {
-        invokeLater(
-            {
-                model.commit()
-            },
-            project
-        )
-    }
+    private val log : Logger = Logger.getInstance(VFSListener::class.java)
 
     private fun doAddSourceFolder(project: Project, virtualFile: VirtualFile, isTestFolder: Boolean) {
         val model = getModelForFile(project, virtualFile)
+
         if (null != model) {
             val contentEntry = getContentEntry(model, virtualFile)
+
             if (null != contentEntry) {
                 val sourceRoots = model.getSourceRoots(if (!isTestFolder) SOURCE else TEST_SOURCE)
+
                 if (!sourceRoots.contains(virtualFile)) {
                     contentEntry.addSourceFolder(virtualFile, isTestFolder)
                     invokeCommit(model, project)
@@ -130,12 +74,16 @@ internal class VFSListener : BulkFileListener {
 
     private fun doMoveSourceFolderAfter(project: Project, templatesPath: String, newParent: VirtualFile, isTestFolder: Boolean) {
         val virtualFile = VfsUtil.findFileByURL(URL("${newParent}/${templatesPath}"))
+
         if (null != virtualFile) {
             val model = getModelForFile(project, virtualFile)
+
             if (null != model) {
                 val contentEntry = getContentEntry(model, virtualFile)
+
                 if (null != contentEntry) {
                     val sourceRoots = model.getSourceRoots(if (!isTestFolder) SOURCE else TEST_SOURCE)
+
                     if (!sourceRoots.contains(virtualFile)) {
                         contentEntry.addSourceFolder(virtualFile, isTestFolder)
                         invokeCommit(model, project)
@@ -153,10 +101,13 @@ internal class VFSListener : BulkFileListener {
 
     private fun doMoveSourceFolderBefore(project: Project, templatesPath: String, oldParent: VirtualFile) {
         val virtualFile = VfsUtil.findFileByURL(URL("${oldParent}/${templatesPath}"))
+
         if (null != virtualFile) {
             val model = getModelForFile(project, virtualFile)
+
             if (null != model) {
                 val contentEntry = getContentEntry(model)
+
                 if (null != contentEntry) {
                     contentEntry
                         .sourceFolders
@@ -177,8 +128,10 @@ internal class VFSListener : BulkFileListener {
 
     private fun doRemoveSourceFolder(project: Project, virtualFile: VirtualFile) {
         val model = getModelForFile(project, virtualFile)
+
         if (null != model) {
             val contentEntry = getContentEntry(model)
+
             if (null != contentEntry) {
                 contentEntry
                     .sourceFolders
@@ -198,12 +151,16 @@ internal class VFSListener : BulkFileListener {
 
     private fun doRenameSourceFolderAfter(project: Project, newPath: String, isTestFolder: Boolean) {
         val virtualFile = VfsUtil.findFileByIoFile(File(newPath), true)
+
         if (null != virtualFile) {
             val model = getModelForFile(project, virtualFile)
+
             if (null != model) {
                 val contentEntry = getContentEntry(model, virtualFile)
+
                 if (null != contentEntry) {
                     val sourceRoots = model.getSourceRoots(if (!isTestFolder) SOURCE else TEST_SOURCE)
+
                     if (!sourceRoots.contains(virtualFile)) {
                         contentEntry.addSourceFolder(virtualFile, isTestFolder)
                         invokeCommit(model, project)
@@ -223,15 +180,17 @@ internal class VFSListener : BulkFileListener {
         val virtualFile = VfsUtil.findFileByIoFile(File(oldPath), true)
         if (null != virtualFile) {
             val model = getModelForFile(project, virtualFile)
+
             if (null != model) {
                 val contentEntry = getContentEntry(model)
+
                 if (null != contentEntry) {
                     contentEntry
                         .sourceFolders
                         .filter { it.file?.equals(virtualFile) ?: false }
                         .forEach { sourceFolder ->
                             contentEntry.removeSourceFolder(sourceFolder)
-                            model.commit()
+                            invokeCommit(model, project)
 
                             log.info(
                                 String.format(